@@ -5,7 +5,7 @@ using System.Net.Http;
 
 namespace NBXplorer
 {
-	public class RPCClientProvider : IRPCClients
+	public class RPCClientProvider
 	{
 		Dictionary<string, RPCClient> _ChainConfigurations = new Dictionary<string, RPCClient>();
 		public RPCClientProvider(ExplorerConfiguration configuration, IHttpClientFactory httpClientFactory)
@@ -15,7 +15,7 @@ namespace NBXplorer
 				var rpc = config?.RPC;
 				if (rpc != null)
 				{
-					rpc.HttpClient = httpClientFactory.CreateClient(nameof(IRPCClients));
+					rpc.HttpClient = httpClientFactory.CreateClient(nameof(RPCClientProvider));
 					_ChainConfigurations.Add(config.CryptoCode, rpc);
 				}
 			}
